Removed Windows compilation warnings

+++ code/src/drivers/MySQL/MySQLStatement.cpp
@@ -565,7 +565,7 @@ bool MySQLStatement::bindVarCharField(MYSQL_BIND& bind, MySQLStatementField* fie
         /// Fetch truncated, enlarge buffer and fetch again
         field->checkSize(dataLength);
         bind.buffer = field->getBuffer();
-        bind.buffer_length = field->bufferSize();
+        bind.buffer_length = (unsigned long) field->bufferSize();
         if (mysql_stmt_fetch_column(statement(), &bind, (unsigned) fieldIndex, 0) != 0)
             throwMySQLError();
         fieldSizeChanged = true;
